@@ -3,34 +3,38 @@
     <div class="row gx-5">
       <div class="col">
         <div class="p-3">
-          <h5><i class="bi bi-clock-history"></i> Yesterday's Stats</h5>
+          <h5>
+            <i class="bi bi-clock-history"></i> Yesterday's Stats
+          </h5>
           <stats v-bind:day="yesterday"></stats>
         </div>
       </div>
       <div class="col">
         <div class="p-3">
-          <h5><i class="bi bi-bar-chart-fill"></i> Today's Stats</h5>
+          <h5>
+            <i class="bi bi-bar-chart-fill"></i> Today's Stats
+          </h5>
           <stats v-bind:day="today"></stats>
         </div>
       </div>
       <div class="col">
         <div class="p-3">
-          <h5><i class="bi bi-speedometer"></i> List</h5>
-          <difficulty v-bind:mode="mode" v-bind:lists="lists" />
+          <h5>
+            <i class="bi bi-speedometer"></i> List
+          </h5>
+          <difficulty v-bind:mode="mode" @modeChange="modeChange" />
         </div>
       </div>
       <div class="col">
         <div class="p-3">
-          <h5><i class="bi bi-github"></i> Source</h5>
+          <h5>
+            <i class="bi bi-github"></i> Source
+          </h5>
           <p>
-            <a href="https://github.com/taiidani/guess-my-word/releases"
-              >Changelog</a
-            >
+            <a href="https://github.com/taiidani/guess-my-word/releases">Changelog</a>
           </p>
           <p>
-            <a href="https://github.com/taiidani/guess-my-word"
-              >View on GitHub</a
-            >
+            <a href="https://github.com/taiidani/guess-my-word">View on GitHub</a>
           </p>
         </div>
       </div>
@@ -41,87 +45,67 @@
 <script>
 import Stats from "./Stats.vue";
 import Difficulty from "./FooterMode.vue";
-
-let yesterday = {
-  word: null,
-  completions: 0,
-  bestRun: 0,
-  avgRun: 0,
-};
-
-let today = {
-  completions: 0,
-  bestRun: 0,
-  avgRun: 0,
-};
-
-let lists = [];
+import { nextTick } from 'vue';
 
 export default {
   components: { Stats, Difficulty },
   name: "Footer",
   props: ["mode"],
+  emits: ["modeChange"],
+  methods: {
+    modeChange: function (newMode) {
+      this.$emit("modeChange", newMode);
+      nextTick(() => { refreshStats.call(this); });
+    }
+  },
   data() {
-    window.setInterval(() => {
-      refreshStats(this.mode);
-    }, 60000);
-    refreshStats(this.mode);
-
-    // Populate the lists dropdown
-    fetch("/api/lists")
-      .then((response) => response.json())
-      .then((data) => {
-        console.debug(data);
-
-        if (data.error) {
-          console.error(data.error);
-          return;
-        }
-
-        data.forEach((item) => {
-          lists.push(item);
-        });
-      })
-      .catch((err) => {
-        console.error(err);
-      });
-
     return {
-      yesterday: yesterday,
-      today: today,
-      lists: lists,
+      yesterday: {
+        word: null,
+        completions: 0,
+        bestRun: 0,
+        avgRun: 0,
+      },
+      today: {
+        completions: 0,
+        bestRun: 0,
+        avgRun: 0,
+      },
     };
   },
+  mounted() {
+    refreshStats.call(this);
+  }
 };
 
-function refreshStats(mode) {
+function refreshStats() {
+  console.debug("Refreshing stats for mode:", this.mode);
   const dt = new Date();
   const params = new URLSearchParams({
     date: Math.floor(dt.getTime() / 1000) - 24 * 60 * 60, // Subtract 1 day
     tz: dt.getTimezoneOffset(),
-    mode: mode,
+    mode: this.mode,
   });
 
   fetch("/api/stats?" + params.toString())
     .then((response) => response.json())
     .then((data) => {
-      console.debug(data);
-
       if (data.error) {
         console.error(data.error);
         return;
       }
 
+      console.debug(data);
       var stats = analyzeStats(data.word.guesses);
-      yesterday.word = data.word.value;
-      yesterday.completions = stats.completions;
-      yesterday.bestRun = stats.bestRun;
-      yesterday.avgRun = stats.avgRun;
+      this.yesterday.word = data.word.value;
+      this.yesterday.completions = stats.completions;
+      this.yesterday.bestRun = stats.bestRun;
+      this.yesterday.avgRun = stats.avgRun;
 
       stats = analyzeStats(data.today.guesses);
-      today.completions = stats.completions;
-      today.bestRun = stats.bestRun;
-      today.avgRun = stats.avgRun;
+      this.today.completions = stats.completions;
+      this.today.bestRun = stats.bestRun;
+      this.today.avgRun = stats.avgRun;
     })
     .catch((err) => {
       console.error(err);
