@@ -3,11 +3,9 @@
     <select id="mode" class="form-select" v-on:change="modeChange">
       <option
         v-for="list in lists"
-        :key="list.name"
-        v-bind:selected="mode == list.name.toLowerCase()"
-      >
-        {{ list.name }}
-      </option>
+        :key="list"
+        v-bind:selected="mode == list.toLowerCase()"
+      >{{ list }}</option>
     </select>
   </div>
 </template>
@@ -15,13 +13,37 @@
 <script>
 export default {
   name: "FooterMode",
-  props: ["mode", "lists"],
-  methods: {
-    modeChange: modeChange,
+  emits: ['modeChange'],
+  props: {
+    mode: String,
   },
-};
+  methods: {
+    modeChange: function (evt) { this.$emit("modeChange", evt.target.value.toLowerCase()); },
+  },
+  data() {
+    return {
+      lists: []
+    }
+  },
+  mounted() {
+    // Populate the lists dropdown
+    fetch("/api/lists")
+      .then((response) => response.json())
+      .then((data) => {
+        console.debug(data);
 
-function modeChange(evt) {
-  window.location.replace("/?mode=" + evt.target.value);
-}
+        if (data.error) {
+          console.error(data.error);
+          return;
+        }
+
+        data.forEach((item) => {
+          this.lists.push(item.name);
+        });
+      })
+      .catch((err) => {
+        console.error(err);
+      });
+  }
+};
 </script>
