@@ -1,13 +1,9 @@
 <template>
   <div class="row">
     <div class="col-md-12">
-      <guess-list
-        name="before"
-        v-bind:list="state.before"
-        v-bind:known="known"
-      />
+      <guess-list name="before" v-bind:list="state.before" v-bind:known="known" />
 
-      <guess-form v-bind:mode="mode" v-bind:state="state" />
+      <guess-form v-bind:mode="mode" v-bind:state="state" @guess="guess" />
 
       <guess-list name="after" v-bind:list="state.after" v-bind:known="known" />
     </div>
@@ -21,16 +17,16 @@ import GuessList from "./GuessList.vue";
 export default {
   name: "Guesser",
   components: { GuessForm, GuessList },
-  props: ["mode"],
+  props: ["mode", "state"],
   data() {
-    return {
-      state: loadState(this.mode),
-    };
+    return {};
+  },
+  methods: {
+    guess: guess
   },
   computed: {
     // The letters currently known
     known: function () {
-      var found = 0;
       if (this.state.after.length == 0 || this.state.before.length == 0) {
         return "";
       }
@@ -49,70 +45,37 @@ export default {
   },
 };
 
-// loadState will restore the browser's state object from the stored sessionStorage state
-function loadState(mode) {
-  let state = {
-    version: 0.9, // Schema version of the state that has been stored
-    before: [], // after tracks all guesses that the correct word is before
-    after: [], // after tracks all guesses that the correct word is after
-    start: new Date(), // start tracks when the guessing started
-    end: null, // end tracks when the correct guess was made
-    idleTime: 0, // idleTime tracks how long the user was blocked from guessing (network time)
-    guesses: 0, // guesses is the number of guesses that the user has made
-    answer: "", // answer stores the correct answer once it has been found
-    save: (state) => {
-      saveState(mode, state);
-    },
-  };
-
-  // Determine if session storage is supported
-  if (typeof Storage === "undefined") {
-    alert(
-      "Your browser does not appear to support the Local Storage API. Please upgrade to a modern browser in order to persist guesses across page reloads."
-    );
-    return state;
-  } else if (typeof sessionStorage["state-" + mode] === "undefined") {
-    // First-time page load, no state
-    console.log("First time page load; initial state");
-    return state;
-  }
-
-  // Load from storage
-  var incomingState = JSON.parse(sessionStorage["state-" + mode]);
-
-  // Massage string dates back into Date objects
-  incomingState.start = new Date(incomingState.start);
-  if (incomingState.end != null) {
-    incomingState.end = new Date(incomingState.end);
-  }
-
-  // Only assign the state if it started on the same day as the current word
-  // Having this logic client-side allows a user to keep guessing their word even after
-  // a new word has been rotated in. As long as they do not refresh their page their
-  // state's "Start" property will lock them to the same day.
-  if (
-    incomingState.start.getDate() == state.start.getDate() &&
-    incomingState.start.getMonth() == state.start.getMonth()
-  ) {
-    state = incomingState;
-  }
-
-  state.save = (state) => {
-    saveState(mode, state);
-  };
-
-  return state;
-}
-
-function saveState(mode, state) {
-  // Determine if session storage is supported
-  if (typeof Storage === "undefined") {
+function guess(data, word) {
+  if (word === null) {
+    console.error("guessed word is: null");
+    return;
+  } else if (word === undefined) {
+    console.error("guessed word is: undefined");
     return;
   }
 
-  const serialized = JSON.stringify(state);
-  sessionStorage["state-" + mode] = serialized;
-  console.debug(serialized);
+  this.state.guesses += 1;
+  if (data.after) {
+    this.state.before.push(word);
+    this.state.before.sort();
+  } else if (data.before) {
+    this.state.after.push(word);
+    this.state.after.sort();
+  } else if (data.correct) {
+    this.state.answer = word;
+    this.state.end = new Date();
+  }
+
+  scrollView();
+  this.state.save(this.state);
+}
+
+function scrollView() {
+  // scroll screen to last after, if available
+  const matches = document.querySelectorAll(".before li:nth-last-child(2)");
+  if (matches.length > 0) {
+    matches[0].scrollIntoView();
+  }
 }
 </script>
 
